requirements.txt
django==5.2
Pillow
mysqlclient
django-unfold
django-paypal